좋아요 기능 완료

+++ BE/src/main/java/com/ssafy/article/controller/ArticleController.java
@@ -6,6 +6,7 @@ import java.util.List;
 import org.springframework.data.repository.query.Param;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -18,6 +19,8 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.ssafy.article.model.dto.ArticleDto;
 import com.ssafy.article.model.service.ArticleService;
+import com.ssafy.member.model.dto.MemberDto;
+import com.ssafy.security.auth.MemberDetails;
 
 @RestController
 @RequestMapping("article")
@@ -50,42 +53,72 @@ public class ArticleController {
 
 	// 게시글 작성
 	@PostMapping("/writeArticle")
-	public ResponseEntity<?> writeArticle(@RequestBody ArticleDto articleDto) {
-		if (service.writeArticle(articleDto) > 0) {
-			return new ResponseEntity<>(HttpStatus.CREATED);
-		} else {
+	public ResponseEntity<?> writeArticle(@RequestBody ArticleDto articleDto, @AuthenticationPrincipal MemberDetails memberDetails) {
+		MemberDto loginMember = memberDetails.getMember();
+		if (articleDto.getArticle_writer().equals(loginMember.getMemberId())) {
+			if (service.writeArticle(articleDto) > 0) {
+				return new ResponseEntity<>(HttpStatus.CREATED);
+			} else {
+				return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+			}
+		}else {
 			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
-		}
+		}	
 	}
 
 	// 공지 작성
 	@PostMapping("/writeNotice")
-	public ResponseEntity<?> writeNotice(@RequestBody ArticleDto articleDto) {
-		if (service.writeNotice(articleDto) > 0) {
-			return new ResponseEntity<>(HttpStatus.CREATED);
-		} else {
+	public ResponseEntity<?> writeNotice(@RequestBody ArticleDto articleDto, @AuthenticationPrincipal MemberDetails memberDetails) {
+		MemberDto loginMember = memberDetails.getMember();
+		if (loginMember.getMemberRole().equals("관리자") && articleDto.getArticle_writer().equals(loginMember.getMemberId())) {
+			if (service.writeNotice(articleDto) > 0) {
+				return new ResponseEntity<>(HttpStatus.CREATED);
+			} else {
+				return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+			}
+		}else {
 			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
 		}
 	}
 
 	// 게시글 수정
 	@PutMapping("/modify")
-	public ResponseEntity<?> modify(@RequestBody ArticleDto articleDto) {
-		if (service.modify(articleDto) > 0) {
-			return new ResponseEntity<>(HttpStatus.OK);
-		} else {
+	public ResponseEntity<?> modify(@RequestBody ArticleDto articleDto, @AuthenticationPrincipal MemberDetails memberDetails) {
+		MemberDto loginMember = memberDetails.getMember();
+		if (articleDto.getArticle_writer().equals(loginMember.getMemberId())) {
+			if (service.modify(articleDto) > 0) {
+				return new ResponseEntity<>(HttpStatus.OK);
+			} else {
+				return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+			}
+		}else {
 			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
 		}
 	}
 
 	// 게시글 삭제
 	@DeleteMapping("/{article_id}")
-	public ResponseEntity<?> remove(@PathVariable("article_id") int articleId) {
-		if (service.remove(articleId) > 0) {
-			return new ResponseEntity<>(HttpStatus.OK);
-		} else {
+	public ResponseEntity<?> remove(@PathVariable("article_id") int articleId, @AuthenticationPrincipal MemberDetails memberDetails) {
+		MemberDto loginMember = memberDetails.getMember();
+		if (service.selectOne(articleId).getArticle_writer().equals(loginMember.getMemberId())) {
+			if (service.remove(articleId) > 0) {
+				return new ResponseEntity<>(HttpStatus.OK);
+			} else {
+				return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+			}
+		}else {
 			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
 		}
 	}
-
+	
+	//게시글 좋아요
+	@GetMapping("/heart/{article_id}")
+	public ResponseEntity<?> heart(@PathVariable("article_id") int articleId, @AuthenticationPrincipal MemberDetails memberDetails){
+		MemberDto loginMember = memberDetails.getMember();
+		if (service.pressHeart(loginMember.getMemberId(), articleId) > 0) {
+			return new ResponseEntity<>(HttpStatus.OK);
+		}else {
+			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+		}
+	}
 }
